% Generated by roxygen2 (4.0.1): do not edit by hand
\name{testNumericDiff}
\alias{testNumericDiff}
\title{Function that checks to see if the difference between 2 number is small enough.}
\usage{
testNumericDiff(queryVals, targetVals, cutoffFraction)
}
\arguments{
\item{queryVals,targetVals}{vectors of values to be compared (pairwise comparison will be performed)}

\item{cutoffFraction}{the value of cutoff to be used to declare if the two numbers are close enough.}
}
\value{
vector of same length as input vectors queryVals and targetVals, recording for each pair of numbers whether they pas the cutoff (TRUE) or not (FALSE).
}
\description{
Function that checks to see if the difference between 2 number is small enough.
}
\details{
for 2 numbers a and b, the function checks to see if |a-b|/min(a,b) <= cutoff.
}
\seealso{
Other unitTestsFunctions: \code{\link{testQuantiles}}
}

